+++ src/main/java/com/centralesupelec/chowchow/TMDB/web/SearchWebController.java
@@ -42,7 +42,7 @@ public class SearchWebController {
       path = "/{id}",
       method = RequestMethod.GET,
       produces = MediaType.APPLICATION_JSON_VALUE)
-  public ResponseEntity findShowById(@PathVariable int id) {
+  public ResponseEntity findShowById(@PathVariable Long id) {
     try {
       TMDBShowDTO search = this.searchController.findShowById(id).getBody();
       return ResponseEntity.status(HttpStatus.OK).body(search);
